fix bug

+++ association/tests/test_association.py
@@ -5,7 +5,7 @@ from mastercontrat import asgi  # noqa
 from mastercontrat import wsgi  # noqa
 
 from django.test import RequestFactory
-
+from django.core.exceptions import ObjectDoesNotExist
 from frontpage.tests.constants import ANONYMOUS
 from association.views import association
 from association.models import MemberOffice
@@ -16,38 +16,58 @@ def test_association():
     request = RequestFactory().get("")
     request.user = ANONYMOUS
 
-    pres = MemberOffice.objects.create(
-        name="testpres",
-        link='formationtest',
-        img_link='2022test',
-        genre='f',
-        promotion='xxx',
-        role='president'
-    )
-    vice = MemberOffice.objects.create(
-        name="testvice",
-        link='formationtest',
-        img_link='2022test',
-        genre='f',
-        promotion='xxx',
-        role='vice'
-    )
-    secre = MemberOffice.objects.create(
-        name="testsecret",
-        link='formationtest',
-        img_link='2022test',
-        genre='f',
-        promotion='xxx',
-        role='secretary'
-    )
-    treasu = MemberOffice.objects.create(
-        name="testtreas",
-        link='formationtest',
-        img_link='2022test',
-        genre='f',
-        promotion='xxx',
-        role='treasurer'
-    )
+    try:
+        MemberOffice.objects.get(role='president')
+        pres = False
+    except ObjectDoesNotExist:
+        pres = MemberOffice.objects.create(
+            name="testpres",
+            link='formationtest',
+            img_link='2022test',
+            genre='f',
+            promotion='xxx',
+            role='president'
+        )
+    
+    try:
+        MemberOffice.objects.get(role='vice')
+        vice = False
+    except ObjectDoesNotExist:
+        vice = MemberOffice.objects.create(
+            name="testvice",
+            link='formationtest',
+            img_link='2022test',
+            genre='f',
+            promotion='xxx',
+            role='vice'
+        )
+    
+    try:
+        MemberOffice.objects.get(role='secretary')
+        secre = False
+    except ObjectDoesNotExist:
+        secre = MemberOffice.objects.create(
+            name="testsecret",
+            link='formationtest',
+            img_link='2022test',
+            genre='f',
+            promotion='xxx',
+            role='secretary'
+        )
+    
+    try:
+        MemberOffice.objects.get(role='treasurer')
+        treasu = False
+    except ObjectDoesNotExist:
+        treasu = MemberOffice.objects.create(
+            name="testtreas",
+            link='formationtest',
+            img_link='2022test',
+            genre='f',
+            promotion='xxx',
+            role='treasurer'
+        )
+
     request = RequestFactory().get("")
     request.user = ANONYMOUS
 
@@ -55,7 +75,11 @@ def test_association():
 
     assert view.status_code == 200
 
-    pres.delete()
-    vice.delete()
-    secre.delete()
-    treasu.delete()
+    if pres:
+        pres.delete()
+    if vice:
+        vice.delete()
+    if secre:
+        secre.delete()
+    if treasu:
+        treasu.delete()
